Move PlatformStatusAction into consensus-utility
