+++ platform-sdk/consensus-pces/src/main/java/org/hiero/consensus/pces/PcesModule.java
@@ -19,8 +19,8 @@ import org.hiero.consensus.metrics.statistics.EventPipelineTracker;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.hashgraph.EventWindow;
 import org.hiero.consensus.model.node.NodeId;
-import org.hiero.consensus.model.status.PlatformStatusAction;
 import org.hiero.consensus.state.signed.ReservedSignedState;
+import org.hiero.consensus.status.actions.PlatformStatusAction;
 
 /**
  * Public interface of the pces module which is responsible for the preconsensus event stream (PCES). It provides
