+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/status/StatusActionSubmitter.java
@@ -2,7 +2,7 @@
 package org.hiero.consensus.status;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
-import org.hiero.consensus.model.status.PlatformStatusAction;
+import org.hiero.consensus.status.actions.PlatformStatusAction;
 
 /**
  * A functional interface for submitting status actions
