+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/status/actions/FallenBehindAction.java
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.consensus.status.actions;
 
-import org.hiero.consensus.model.status.PlatformStatusAction;
-
 /**
  * An action that is triggered when the platform falls behind.
  */
