+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/status/actions/FreezePeriodEnteredAction.java
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.consensus.status.actions;
 
-import org.hiero.consensus.model.status.PlatformStatusAction;
-
 /**
  * An action to indicate that consensus has advanced past a freeze timestamp
  *
