+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/status/actions/ReconnectCompleteAction.java
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.consensus.status.actions;
 
-import org.hiero.consensus.model.status.PlatformStatusAction;
-
 /**
  * An action to indicate that the reconnect process has completed
  *
