+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/status/actions/StartedReplayingEventsAction.java
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.consensus.status.actions;
 
-import org.hiero.consensus.model.status.PlatformStatusAction;
-
 /**
  * An action that is triggered when the platform starts replaying events from the preconsensus event stream.
  */
