+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/system/DefaultPlatformMonitor.java
@@ -17,9 +17,9 @@ import org.hiero.consensus.model.notification.IssNotification.IssType;
 import org.hiero.consensus.model.quiescence.QuiescenceCommand;
 import org.hiero.consensus.model.state.StateSavingResult;
 import org.hiero.consensus.model.status.PlatformStatus;
-import org.hiero.consensus.model.status.PlatformStatusAction;
 import org.hiero.consensus.status.StatusStateMachine;
 import org.hiero.consensus.status.actions.CatastrophicFailureAction;
+import org.hiero.consensus.status.actions.PlatformStatusAction;
 import org.hiero.consensus.status.actions.SelfEventReachedConsensusAction;
 import org.hiero.consensus.status.actions.StateWrittenToDiskAction;
 import org.hiero.consensus.status.actions.TimeElapsedAction;
